requirements.txt
ddgs>=9.0
trafilatura>=1.6
google-generativeai>=0.3.0
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
reportlab>=4.0
jinja2>=3.1
